Rework of scoreboard system + hammer behaviour + simplification game process with event + import VR keyboard

+++ Assets/Scripts/Process/Gun Club/Game/DataGunClubWave.cs
@@ -8,7 +8,7 @@ namespace Assets.Scripts.Process.GunClub.Game
     {
         public int NumberOfTarget;
         public float CooldownSpawn;
-        public float CooldownAlive;
+        public int CooldownAlive;
         public List<GameObject> Prefab;
 
     }
